break into component
